+++ client/CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to the "Harbour and xHarbour" extension will be documented in this file.
 
 # 1.0.6
+- **Server** better classData, classVar, classMethod support
  - **Syntax** better classData, classVar, classMethod support
  - **Debugger** better handshake